feat: integrating employee flows into employee namespace
