+++ src/components/Employee/SelfOnboardingFlow/selfOnboardingMachine.ts
@@ -1,12 +1,7 @@
 import { transition, reduce, state, invoke, createMachine } from 'robot3'
+import type { SelfOnboardingContextInterface } from './SelfOnboardingComponents'
+import { Profile, Taxes, PaymentMethod, OnboardingSummary } from './SelfOnboardingComponents'
 import { componentEvents } from '@/shared/constants'
-import {
-  Profile,
-  Taxes,
-  PaymentMethod,
-  OnboardingSummary,
-} from '@/components/Flow/EmployeeSelfOnboardingFlow/EmployeeSelfOnboardingComponents'
-import type { EmployeeSelfOnboardingContextInterface } from '@/components/Flow/EmployeeSelfOnboardingFlow/EmployeeSelfOnboardingFlow'
 import type { DocumentSignerContextInterface } from '@/components/Employee/DocumentSigner/documentSignerStateMachine'
 import { DocumentListContextual } from '@/components/Employee/DocumentSigner/documentSignerStateMachine'
 import { documentSignerMachine } from '@/components/Employee/DocumentSigner/stateMachine'
@@ -26,7 +21,7 @@ export const employeeSelfOnboardingMachine = {
       componentEvents.EMPLOYEE_SELF_ONBOARDING_START,
       'employeeProfile',
       reduce(
-        (ctx: EmployeeSelfOnboardingContextInterface): EmployeeSelfOnboardingContextInterface => ({
+        (ctx: SelfOnboardingContextInterface): SelfOnboardingContextInterface => ({
           ...ctx,
           component: Profile,
         }),
@@ -38,7 +33,7 @@ export const employeeSelfOnboardingMachine = {
       componentEvents.EMPLOYEE_PROFILE_DONE,
       'employeeTaxes',
       reduce(
-        (ctx: EmployeeSelfOnboardingContextInterface): EmployeeSelfOnboardingContextInterface => ({
+        (ctx: SelfOnboardingContextInterface): SelfOnboardingContextInterface => ({
           ...ctx,
           component: Taxes,
         }),
@@ -49,7 +44,7 @@ export const employeeSelfOnboardingMachine = {
     transition(
       componentEvents.EMPLOYEE_TAXES_DONE,
       'employeePaymentMethod',
-      reduce((ctx: EmployeeSelfOnboardingContextInterface) => ({
+      reduce((ctx: SelfOnboardingContextInterface) => ({
         ...ctx,
         component: PaymentMethod,
       })),
@@ -59,7 +54,7 @@ export const employeeSelfOnboardingMachine = {
     transition(
       componentEvents.EMPLOYEE_PAYMENT_METHOD_DONE,
       'employeeDocumentSigner',
-      reduce((ctx: EmployeeSelfOnboardingContextInterface) => ({
+      reduce((ctx: SelfOnboardingContextInterface) => ({
         ...ctx,
       })),
     ),
@@ -70,7 +65,7 @@ export const employeeSelfOnboardingMachine = {
     transition(
       componentEvents.ROBOT_MACHINE_DONE,
       'index',
-      reduce((ctx: EmployeeSelfOnboardingContextInterface) => ({
+      reduce((ctx: SelfOnboardingContextInterface) => ({
         ...ctx,
         component: OnboardingSummary,
       })),
