+++ src/components/Employee/index.ts
@@ -7,3 +7,5 @@ export { Taxes } from './Taxes'
 export { PaymentMethod } from './PaymentMethod'
 export { Landing } from './Landing'
 export { DocumentSignerFlow } from './DocumentSigner'
+export { OnboardingFlow } from './OnboardingFlow/OnboardingFlow'
+export { SelfOnboardingFlow } from './SelfOnboardingFlow/SelfOnboardingFlow'
