--- src/components/Flow/StateMachines/index.ts
@@ -1,3 +1,0 @@
-export * from './employeeOnboarding'
-export * from './employeeSelfOnboarding'
-// export * from './companyOnboarding'
