--- src/components/Flow/index.ts
@@ -1,2 +1,0 @@
-export { EmployeeOnboardingFlow } from './EmployeeOnboardingFlow'
-export { EmployeeSelfOnboardingFlow } from './EmployeeSelfOnboardingFlow/EmployeeSelfOnboardingFlow'
