+++ src/components/index.ts
@@ -1,4 +1,3 @@
 // Explicit exports available to partners
 export * as Company from './Company'
 export * as Employee from './Employee'
-export * as Flows from './Flow'
