+++ src/index.ts
@@ -1,6 +1,3 @@
 export * from '@/components'
 export * from '@/contexts'
 export { componentEvents } from '@/shared/constants'
-export { EmployeeOnboardingFlow } from '@/components/Flow/EmployeeOnboardingFlow'
-
-export { EmployeeSelfOnboardingFlow } from '@/components/Flow/EmployeeSelfOnboardingFlow/EmployeeSelfOnboardingFlow'
